split gui into sep. file
